Setup/fix test
